Simplify page count algorithm
Rather than replaying the entire Status table, we now just find the
difference between the first and last updates for a given printer.

This is adequate for now because none of the printers have reset
their meters yet, but we'll need to have an "intelligent" replay
algorithm that can detect resets in the future.