Add handle_status_update stored procedure to schema
Eventually this will completely obsolete the Outages table, since it
has the same information plus the periods of availability. It's possible
to implement this as a view, but in practice there's no easy way to get
good performance without materializing the view (which is impossible
in Postgres) so this is a hack around that. A working, pretty hack, though.